@@ -1,5 +1,5 @@
 import { FormControl, FormArray } from '@angular/forms';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 
 import { BookStoreService } from '../../shared/book-store.service';
